pyproject.toml
[tool.poetry]
name = "pinboard-queue"
version = "0.1.0"
description = ""
authors = ["Simon Lundström <github-commits@soy.se>"]
include = ["pinboard_post.capnp"]

[tool.poetry.dependencies]
python = "^3.7"
pika = "^1.2.0"
pycapnp = "^1.0.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.790"
pylint = "^2.6.0"
black = "^20.8b1"

[tool.poetry.scripts]
pinboard-queue = 'pinboard_queue:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
